refactor(ui): reorganize store structure and update imports for StoreInitializer and UI store, remove any types
