+++ ui/components/providers/index.ts
@@ -1,3 +1,4 @@
+export * from "../../store/ui/store-initializer";
 export * from "./add-provider-button";
 export * from "./credentials-update-info";
 export * from "./forms/delete-form";
@@ -5,4 +6,3 @@ export * from "./link-to-scans";
 export * from "./muted-findings-config-button";
 export * from "./provider-info";
 export * from "./radio-group-provider";
-export * from "./store-initializer";
