+++ ui/store/index.ts
@@ -1,1 +1,1 @@
-export * from "./ui/ui-store";
+export * from "./ui/store";
